+++ shared/src/commonMain/kotlin/io/github/gatrongdev/kbignum/math/KBigNumberExtensions.kt
@@ -95,7 +95,7 @@ operator fun KBigDecimal.times(other: KBigDecimal): KBigDecimal = this.multiply(
  * @return The quotient with scale 10 and HALF_UP rounding
  * @throws ArithmeticException if other is zero
  */
-operator fun KBigDecimal.div(other: KBigDecimal): KBigDecimal = this.divide(other, 10, 4)
+operator fun KBigDecimal.div(other: KBigDecimal): KBigDecimal = this.divide(other, 10, KBRoundingMode.HalfUp)
 
 /**
  * Unary minus operator for KBigDecimal.